Extract ReactionLogger.
